extracted skybox logic from 3rd/world/index
